pet-clinic-data/src/test/java/akshitone/springtutorial/petclinicthird/services/map/OwnerServiceMapTest.java
package akshitone.springtutorial.petclinicthird.services.map;

import akshitone.springtutorial.petclinicthird.model.Owner;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

import java.util.Set;

class OwnerServiceMapTest {
    OwnerServiceMap ownerServiceMap;

    final Long ownerId = 1L;
    final String ownerLastName = "Johnson";

    @BeforeEach
    void setUp() {
        ownerServiceMap = new OwnerServiceMap(new PetTypeServiceMap(), new PetServiceMap());
        ownerServiceMap.save(Owner.builder().id(ownerId).lastName(ownerLastName).build());
    }

    @Test
    void findAll() {
        Set<Owner> owners = ownerServiceMap.findAll();
        assertEquals(1,owners.size());
    }

    @Test
    void findById() {
        Owner owner = ownerServiceMap.findById(ownerId);
        assertEquals(ownerId,owner.getId());
    }

    @Test
    void saveExistingId() {
        Long ownerId = 2L;
        Owner owner = Owner.builder().id(ownerId).build();
        Owner savedOwner = ownerServiceMap.save(owner);
        assertEquals(ownerId,savedOwner.getId());
    }

    @Test
    void saveNoId() {
        Owner savedOwner = ownerServiceMap.save(Owner.builder().build());
        assertNotNull(savedOwner);
        assertNotNull(savedOwner.getId());
    }

    @Test
    void delete() {
        ownerServiceMap.delete(ownerServiceMap.findById(ownerId));
        assertEquals(0,ownerServiceMap.findAll().size());
    }

    @Test
    void deleteById() {
        ownerServiceMap.deleteById(ownerId);
        assertEquals(0,ownerServiceMap.findAll().size());
    }

    @Test
    void findByLastName() {
        Owner owner = ownerServiceMap.findByLastName(ownerLastName);
        assertNotNull(owner);
        assertEquals(ownerId,owner.getId());
    }
}